pyproject.toml
[tool.poetry]
name = "fastapi-langraph"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "fastapi_langraph"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.111.0"}
uvicorn = "^0.29.0"
pydantic = "^2.6.4"
pydantic-settings = "^2.2.1"
loguru = "^0.7.2"
langgraph = "^0.0.38"
langchain_openai = "^0.1.3"
langchain = "^0.1.16"
bs4 = "^0.0.2"
tavily-python = "^0.3.3"
duckduckgo-search = "^5.3.1"
langchain-community = "^0.0.34"
python-dotenv = "^1.1.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pytest-cov = "^5.0.0"
httpx = "^0.27.0"
pre-commit = "^3.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
